+++ packages/main/src/GroupHeaderListItem.js
@@ -1,4 +1,4 @@
-import Core from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/Core";
+import Bootstrap from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/Bootstrap";
 import ShadowDOM from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/compatibility/ShadowDOM";
 import ListItemBase from "./ListItemBase";
 // Template
@@ -60,7 +60,7 @@ class GroupHeaderListItem extends ListItemBase {
 	}
 }
 
-Core.boot().then(_ => {
+Bootstrap.boot().then(_ => {
 	GroupHeaderListItem.define();
 });
