+++ packages/main/src/Input.js
@@ -1,5 +1,5 @@
 import WebComponent from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/WebComponent";
-import Core from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/Core";
+import Bootstrap from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/Bootstrap";
 import ValueState from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/types/ValueState";
 import ShadowDOM from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/compatibility/ShadowDOM";
 import Icon from "./Icon";
@@ -478,7 +478,7 @@ class Input extends WebComponent {
 	onClose() {}
 }
 
-Core.boot().then(_ => {
+Bootstrap.boot().then(_ => {
 	Input.define();
 });
