+++ packages/main/src/Select.js
@@ -1,5 +1,5 @@
 import WebComponent from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/WebComponent";
-import Core from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/Core";
+import Bootstrap from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/Bootstrap";
 import ShadowDOM from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/compatibility/ShadowDOM";
 import KeyCodes from "@ui5/webcomponents-core/dist/sap/ui/events/KeyCodes";
 import ValueState from "@ui5/webcomponents-base/src/sap/ui/webcomponents/base/types/ValueState";
@@ -325,7 +325,7 @@ class Select extends WebComponent {
 
 Select.ARROWS = [KeyCodes.ARROW_DOWN, KeyCodes.ARROW_UP];
 
-Core.boot().then(_ => {
+Bootstrap.boot().then(_ => {
 	Select.define();
 });
